Commonize ritual data storage into a common base class that DKGStorage can extend. Eventually, signing will also do the same.
